Split fields.py into 4 files
